finalize the layout of the drawer menu
